Time: 57 ms (9.07%) | Memory: 46.6 MB (13.10%) - LeetSync

+++ leetcode/264-ugly-number-ii/ugly-number-ii.java
@@ -19,7 +19,7 @@ class Solution {
                 set.add(num*5);
             }
         }
-    int val = (int)((long) heap.poll());
+    int val =  heap.poll().intValue();
         return val;
 
     }
